requirements.txt
ollama
python-dotenv
google.generativeai
langchain
langchain-google-genai
pydantic